+++ src/os/shared/src/osapi-time.c
@@ -294,7 +294,7 @@ int32 OS_TimerCreate(osal_id_t *timer_id, const char *timer_name, uint32 *accura
     }
     else
     {
-        *accuracy = OS_SharedGlobalVars.MicroSecPerTick;
+        *accuracy = (uint32)OS_SharedGlobalVars.MicroSecPerTick;
     }
 
     return return_code;
